extract to file and add logs
